Refactor content_filtering
