+++ crates/fuzzy_filter/src/content_filtering.rs
@@ -1,22 +1,7 @@
+use crate::{fuzzy_indices_skim, ScorerOutput};
 use extracted_fzy::match_and_score_with_positions;
-pub use fuzzy_matcher::skim::fuzzy_indices as fuzzy_indices_skim;
 use lazy_static::lazy_static;
 use regex::Regex;
-use structopt::clap::arg_enum;
-
-// Implement arg_enum so that we could control it from the command line.
-arg_enum! {
-  #[derive(Debug, Clone)]
-  pub enum ContentFiltering {
-      Full,
-      FileNameOnly,
-      GrepExcludeFilePath,
-  }
-}
-
-// Returns the score and indices of matched chars
-// when the line is matched given the query,
-type ScorerOutput = Option<(i64, Vec<usize>)>;
 
 lazy_static! {
     // match the file path and line number of grep line.
@@ -25,7 +10,7 @@ lazy_static! {
 
 /// Make the arguments order same to Skim's `fuzzy_indices()`.
 #[inline]
-pub(super) fn fuzzy_indices_fzy(line: &str, query: &str) -> ScorerOutput {
+pub fn fuzzy_indices_fzy(line: &str, query: &str) -> ScorerOutput {
     match_and_score_with_positions(query, line).map(|(score, indices)| (score as i64, indices))
 }
